+++ src/main/Main.java
@@ -54,7 +54,7 @@ public class Main extends Application {
         while(scanner.hasNextLine()){
             String line = scanner.nextLine();
             String[] args = line.split(",");
-            modulePQ.addTimeSlot(args[0], Long.parseLong(args[1]), args[2]);
+            modulePQ.addTimeSlot(args[0], Integer.parseInt(args[1]), args[2]);
             moduleSet.add(args[0]);
         }
         scanner.close();
